Sphinx documentation (#78)
* Fix merge (#46)

* Fix Merge

* Updated requirements.txt

Updated the requirements for the project.

* Fixed README (#68)

Added more info for the avatar class in the README.

* Starting to document

Starting to add sphinx documentation.

* Created rst files

Created the rst and html files for every file.

* Testing

Testing different things to make the formatting make sense.

* Fixing mistake

Fixed a mistake from a merge conflict.

* Item Docs

Finished the item docs for sphinx.

* Map Directory

Fixed documentation in Map directory. Will need to find solution for visuals.

* Added Stations

Added station docs. May need to expand on their explanations.

* Fixed Formatting

Fixed the formatting for GameBoard docs and README. Thanks Jean.

* Reformatting

Reformatted and made small changes to some files

* Game Package Classes

Fixed docs for Game package classes like enums and avatar

* Test File Docs

Formatted the Test File docs. Will eventually add proper docstrings to the individual files themselves.

* Formatting

This should be the final formats for the rst files.

* Controller Docs

Made changes for the Controller documents

* Test Docs

Made documentation for the tests.

* Test Docs 2

Last bits of the documentation for the test suit files.

* Utils Docs

Started documentation for the Utils files and the last few that are missing

* Changes

Changes for some files and an update to requirements.txt.

* Extra docs

Should be done adding documentation for everything.

* Fixes

Fixing warnings

* Removal

Removed unnecessary files

* Finalized

Should be done with documentation for main engine.

---------

Co-authored-by: Jean A. Eckelberg <fyrelynx11@gmail.com>