Client rewrite (#131)
* wip

* wip

* Done

Should be done with the start of the utils

* wip

* wip

* changed to run schema without group run

* fixes to unit tests (#124)

changes:
- test_submission.py json
- test_leaderboard.py json
- test_group_run.py json
- run_schema_wo_group_run.py import
- group_teams_schema.py imports

deleted:
- test_run.py
- test database

* changes from last time

* taking in group_run_id for get leaderboard

* main file fixes

* Main file changes

Added some arguments to the main.py file

* Added Leaderboard Builder

Added methods to help collect the info needed for building the leaderboard

* GR command

Added a new command to help with testing for the future.

* fixing client rewrite

* Client testing

Started testing code for the client methods. Will need to revisit soon to fix the errors that are occurring.

* Registration works now

* Passed seed into generation

* Large updates

Finding solutions to client problems and testing the client utils.

* Update crud_submission.py (#130)

added:
- get _latest_submission method
    - gets all teams latest submissions
        - runs with joining on submission_id

* Client utils update

Added getting a sum of a team's score

* Client Runner progress

Made progress on the client runner. Will need to fix some errors

* client_runner fixes

* fixes to visualizer_runner

* updated visualizer_loop

* work on client and visualizer runners

* updated client and runners

* fix to saving video

changed in main.py:
- self.playback_speed = 10 to self.playback_speed = 10.0

* Small changes

Removed comments and unused code

* Updates

Added a small null handling and added a message in the external runner

* Continued work on Client Runner and config file for Client Runner

* fixed tournament id error

* migrated to postgres

* Server update

Tested the switch to postgres and removed the pygame support text

* Client registration and submission fix

Tested and fixed client registration and submission to now show traceback error messages and show exactly what went wrong

* migrated back to sqlite

* fixed some issues

fixed:
- allowing numbers less than 1 for playback speed in visualizer

deleted:
- -server argument in visualizer (did nothing now)

* Updated gitignore

Added temp folders to gitignore

* Used decorators and Result struct

* Added Result Example

* Changes by moody

* error fixing

* Team Type update

Updated the team types in the client runner. Updated gitignore

* Eligible update

Fixed bool error

* fixes some leaderboard stuff

* fixed team register problem

fixed issue with when registering a team with the same name as a pre-existing team name, that the error provided would return that team's vID (no good)

* fixed details and runs in stats

now returns that submission has no details/runs when the list of details/runs for a submission is zero
(will need some more testing)

* Leaderboard commands

Fixed some leaderboard commands to work properly and give proper error messages

* Visualizer bat update

Removed -server in the bat command. Updated gitignore to properly ignore temp folders

* Update .gitignore

Added out.mp4 to the gitignore

* Visualizer update

Default speed for visualizer runner is 2x speed. Changed '... leaderboard -leaderboard_id' to '... leaderboard -id'

* Fixed local time to utc time

* Deleted unused

Deleted unused code that was from the original database

* Sped up tournament list retrieval

---------

Co-authored-by: JuliaCaesar <jmmotzko@gmail.com>
Co-authored-by: Gunnar Moody <115500822+MoodyMan04@users.noreply.github.com>
Co-authored-by: Jean A. Eckelberg <fyrelynx11@gmail.com>
Co-authored-by: Gunnar Moody <moodygman04@gmail.com>
Co-authored-by: CarseKeeper <125390842+CarseKeeper@users.noreply.github.com>